pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agent-company"
version = "0.1.0"
description = "多AI协作讨论与执行框架 - AI Virtual R&D Organization"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }

dependencies = [
    "pydantic>=2.0",
    "aiosqlite>=0.19.0",
    "openai>=1.0",
    "typer[all]>=0.9.0",
    "rich>=13.0",
    "gitpython>=3.1.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
gemini = ["google-generativeai>=0.3.0"]
claude = ["anthropic>=0.18.0"]
web = ["fastapi>=0.100.0", "uvicorn>=0.23.0", "httpx>=0.24.0"]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
]
all = ["agent-company[gemini,claude,web,dev]"]

[project.scripts]
agent-company = "agent_company.cli.app:app"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
